implemented jwt and protected routes

+++ backend/Controllers/goalControllers.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler')
-const Goal=require('../models/goalModel')
+const Goal = require('../models/goalModel')
+const User=require('../models/userModel')
 
 // @desc Get Goals
 // @route GET /api/goals
 // @access private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals=await Goal.find()
+  const goals=await Goal.find( {user:req.user.id})
 
   res.status(200).json(goals)
 })
@@ -20,7 +21,8 @@ const addGoal = asyncHandler(async (req, res) => {
   }
 
   const goal = await Goal.create({
-    text:req.body.text
+    text: req.body.text,
+    user:req.user.id
   })
 
   res.status(200).json(goal)
@@ -36,7 +38,23 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400)
     throw new Error('Goal not found')
   }
-  const updatedGoal= await Goal.findByIdAndUpdate(req.params.id,req.body,{new:true})
+
+  const user = await User.findById(req.user.id)
+  
+  //check for user
+  if (!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+// Make sure the logged in user matches the goal user
+  if (goal.user.toString()!==user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
+  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true })
+  
   res.status(200).json(updatedGoal)
 }
 )
@@ -49,6 +67,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400)
     throw new Error('Goal not found')
+  }
+
+  const user = await User.findById(req.user.id)
+  
+  //check for user
+  if (!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+// Make sure the logged in user matches the goal user
+  if (goal.user.toString()!==user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   await goal.deleteOne()
